feat: P2022-1618 Add card API
